fix build bug

+++ src/lib.bash
@@ -117,14 +117,14 @@ _last_version() {
 }
 
 _downlock() {
-    local prefix=${1##*/} suffix swp=$$$RANDOM$RANDOM;
+    local prefix=${1##*/} suffix swp=$$$RANDOM.$RANDOM;
     suffix=${prefix##*[0-9]};
     [ "$suffix" ] || suffix=".${prefix##*[0-9]\.}";
     [ "${suffix:0:1}" != "." ] && suffix=".${suffix#*.}";
     prefix=${prefix%%-*};
     [ "$prefix" == "${1##*/}" ] && prefix="${prefix%%[0-9]*}";
     printf " ----------- download $prefix ---------------------\n";
-    curl -L --retry 10 -o $TMP/$swp $1 || {
+    [ -f "$TMP/$prefix$suffix" ] || curl -L --retry 10 -o $TMP/$swp $1 || {
         rm -f $TMP/$swp;
         printf "[ERROR] download $prefix fail.\n" >&2;
         return 1
